+++ jstarcraft-ai-lsh/src/main/java/be/tarsos/lsh/families/CosineDistance.java
@@ -47,9 +47,9 @@ import be.tarsos.lsh.Vector;
 public class CosineDistance implements DistanceMeasure {
 
     @Override
-    public double distance(Vector one, Vector other) {
-        double distance = 0;
-        double similarity = one.dot(other) / Math.sqrt(one.dot(one) * other.dot(other));
+    public float distance(Vector one, Vector other) {
+        float distance = 0F;
+        float similarity = one.dot(other) / (float) Math.sqrt(one.dot(one) * other.dot(other));
         distance = 1 - similarity;
         return distance;
     }
